BUG/MINOR: re-enable quick ACK when the connection is established
We've been wondering for a long time why data rate was lower with -F,
it's because quick-ack remains enabled, so ACKs for incoming data are
delayed and the sender has to keep more data in its buffers, causing
send() calls to be smaller.

We now re-enable quick-ack once the connection is established to
avoid this.