MEDIUM: implement MSG_TRUNC on the receive path
Using MSG_TRUNC on linux 2.4 and above discards incoming data up to the
requested amount without copying them. Tests show that doing so allows
us to jump from 29.3 to 40.2 Gbps on the loopback using recv(), compared to
38.6 for splice(). The real advantage is that it's done in a single syscall
versus 3 for splice().

On an atom, with data fetched over the network, the savings are even more
impressive, with CPU dropping from 75-86% to 35-40% at 1 Gbps. It's worth
noting that the same machine shows 98% CPU using splice.

Since this will not be supported on all platforms, we pass NULL in the
buffer's pointer so that others immediately fail instead of corrupting
memory.