MEDIUM: replace the fast close with sending of an RST after receiving the FIN
The fast close (-c) never was really usable, most often it broke and it was not
really compliant with most servers. However we can still disable lingering upon
receipt of the last segment. Doing so increases hit rate by about 5% over the
network but has no effect on a local machine.